refactor(cast): share code fetcher and result printers across commands
